rework: Rename lwjgl2_methods_injector into methods_injector_agent
