--- app_pojavlauncher/src/main/assets/components/lwjgl2_methods_injector/version
@@ -1,1 +1,0 @@
-59b2f04b91fb2c8c20148554f1388007c2265bd8
