+++ methods_injector_agent/src/main/java/org/angelauramc/methodsInjectorAgent/lwjgl2_methods_injector/ALC10Injector.java
@@ -1,4 +1,4 @@
-package org.angelauramc.lwjgl2_methods_injector;
+package org.angelauramc.methodsInjectorAgent.lwjgl2_methods_injector;
 
 import static org.objectweb.asm.Opcodes.*;
